Added a watchpoint implementation that polls; I was getting a very occasional crash that was due to occasional u16 laying in memory in such a way as to look like a valid heap pointer. I need to go back and do ptr scanning for everything, but am thinking about whether it is more automatable to produce the ptr maps, e.g., via lib clang. Until then, I've made some of the data types wide.

+++ include/con4m/datatypes/vm.h
@@ -351,7 +351,7 @@ typedef enum : uint8_t {
 typedef struct {
     c4m_zop_t   op;
     uint8_t     pad;
-    int16_t     module_id;
+    int32_t     module_id;
     int32_t     line_no;
     int32_t     arg;
     int64_t     immediate;
@@ -364,7 +364,7 @@ typedef struct {
     int64_t     impl;
     int64_t     nameoffset;
     c4m_type_t *tid;
-    int16_t     mid;
+    int32_t     mid;
     bool        ffi;
 } c4m_zcallback_t;
 
@@ -468,7 +468,7 @@ typedef struct {
 
 typedef struct {
     uint64_t     zero_magic;
-    uint16_t     zc_object_vers;
+    uint32_t     zc_object_vers;
     c4m_buf_t   *static_data;
     c4m_buf_t   *marshaled_consts;
     int32_t      num_const_objs;
